Generate airbins and add lit
